move fill_constant_batch_size_like api
