+++ python/paddle/fluid/tests/unittests/auto_parallel_gpt_model.py
@@ -18,7 +18,6 @@ import paddle
 import paddle.nn.functional as F
 from paddle import nn, tensor
 from paddle.distributed.fleet import auto
-from paddle.fluid import layers
 from paddle.nn.layer.transformer import _convert_param_attr_to_list
 
 paddle.enable_static()
@@ -212,13 +211,13 @@ class MultiHeadAttention(nn.Layer):
             k, v = self.compute_kv(key, value)
             return self.StaticCache(k, v)
         elif value is None:  # incremental_state
-            k = layers.fill_constant_batch_size_like(
+            k = paddle.tensor.fill_constant_batch_size_like(
                 input=key,
                 shape=[-1, self.num_heads, 0, self.head_dim],
                 dtype=key.dtype,
                 value=0,
             )
-            v = layers.fill_constant_batch_size_like(
+            v = paddle.tensor.fill_constant_batch_size_like(
                 input=key,
                 shape=[-1, self.num_heads, 0, self.head_dim],
                 dtype=key.dtype,
