+++ python/paddle/tensor/__init__.py
@@ -27,6 +27,7 @@ from .creation import diagflat  # noqa: F401
 from .creation import eye  # noqa: F401
 from .creation import linspace  # noqa: F401
 from .creation import fill_constant  # noqa: F401
+from .creation import fill_constant_batch_size_like  # noqa: F401
 from .creation import ones  # noqa: F401
 from .creation import ones_like  # noqa: F401
 from .creation import zeros  # noqa: F401
